% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aagi_colours}
\alias{aagi_colours}
\title{A Vector of the official AAGI Colour Hex Codes.  Other colour palettes are
provided in \pkg{theme.aagi} for plotting purposes but are not considered
to be "official" \acronym{AAGI} colours.}
\format{
A \code{vector} with three (6) colours as hexadecimal values:
\describe{
\item{AAGI AAGI Black}{#414042}
\item{AAGI AAGI Bright Green}{#B6D438}
\item{AAGI AAGI Teal}{#00808B}
\item{AAGI Blue}{#648FD2}
\item{AAGI Green}{#54921E}
\item{AAGI Yellow}{#FFBC42}
}
}
\usage{
aagi_colours
}
\description{
A Vector of the official AAGI Colour Hex Codes.  Other colour palettes are
provided in \pkg{theme.aagi} for plotting purposes but are not considered
to be "official" \acronym{AAGI} colours.
}
\concept{data}
\keyword{datasets}
