% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_pal_info.R
\name{aagi_pal_info}
\alias{aagi_pal_info}
\title{Information About AAGI Palettes}
\usage{
aagi_pal_info()
}
\value{
a \code{data.frame} with two columns, \dQuote{category} and
\dQuote{max_colours} with palette names as row names.
}
\description{
Provides information on \acronym{AAGI} palettes including whether they are
diverging, \dQuote{div} or sequential, \dQuote{seq} and the max number of
colour steps available.
}
\examples{
aagi_pal_info()

}
