% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{aagi_palette}
\alias{aagi_palette}
\title{Sequential, Diverging and Qualitative Colour Palettes for AAGI Publications}
\usage{
aagi_palette(
  n,
  name = c("aagi_BrYl", "aagi_BuOr", "aagi_BuYl", "aagi_GnYl", "aagi_RdBu", "aagi_RdTl",
    "aagi_RdYl", "aagi_TlGn", "aagi_TlYl", "aagi_TlGnYl", "aagi_blues",
    "aagi_bright_greens", "aagi_greens", "aagi_greys", "aagi_reds", "aagi_teals",
    "aagi_oranges", "aagi_yellows", "Okabe_and_Ito"),
  direction = 1
)
}
\arguments{
\item{n}{\code{Integer} Number of different colours in the palette, minimum 3,
maximum 11, depending on palette}

\item{name}{\code{Character} A string value with the name of the desired palette.}

\item{direction}{\code{Integer} Sets the order of colours in the scale.  If \code{1},
the default order is used. If \code{-1}, the order of colours is reversed.}
}
\value{
A vector of hexadecimal colour codes
}
\description{
\acronym{AAGI} palettes for use in charts and plots based on official
\acronym{AAGI} colours and interpolated to provide more colours for use in
maps and other graphical outputs.  These palettes are intended to be
colour-blind friendly, while still being based on and include official
\acronym{AAGI} colours but not all will be official \acronym{AAGI} colours.
They are are palettes that are intended to be useful for creating plots
that incorporate come official  \acronym{AAGI} colours but they may not all
be official \acronym{AAGI} colours.
}
\section{Methods}{
The colours were taken from the \acronym{AAGI} communications style guide
colour palette. In some cases, \emph{e.g.}, \code{aagi_red}, \verb{#8b0b00} value,
did not exist in the official palette and was generated by taking the
complimentary colour of the \code{aagi_teal}, \verb{#00808b} value. The darkest and
lightest values used in interpolating the colours were selected from the
"Tint Colour Variation" list of this website,
\url{https://www.colorhexa.com/c5856d}, using the \acronym{AAGI}
hexadecimal colour code as the input. The interpolated values were all
generated using \url{http://www.vis4.net/palettes/}, ensuring that all
palettes are colourblind safe.
}

\examples{
# sequential blues, max of 7
aagi_palette(n = 7, name = "aagi_blues", direction = 1)

# diverging red teal colours with 11 stops
aagi_palette(n = 11, name = "aagi_RdTl", direction = 1)

}
