% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palette_info.R
\name{aagi_palette_info}
\alias{aagi_palette_info}
\title{Information About AAGI Palettes}
\usage{
aagi_palette_info()
}
\value{
a \code{data.frame} with two columns, \dQuote{category} and
\dQuote{max_colours} with palette names as row names.
}
\description{
Provides information on \acronym{AAGI} palettes including whether they are
diverging, \dQuote{div} or sequential, \dQuote{seq} and the max number of
colour steps available.  \acronym{AAGI} palettes are palettes that are
useful for creating plots that incorporate come official  \acronym{AAGI}
colours but they may not all be official \acronym{AAGI} colours.
}
\examples{
aagi_palette_info()

}
