% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{aagi_palettes}
\alias{aagi_palettes}
\title{List AAGI Palette Colour Combinations}
\usage{
aagi_palettes()
}
\value{
A \code{list} object of \acronym{AAGI} colours grouped by palette
combinations
}
\description{
Returns a list of curated palettes of \acronym{AAGI} colours, which includes:
\describe{
\item{primary}{The three primary \acronym{AAGI} colours}
\item{secondary}{The three secondary support colours in the \acronym{AAGI}
colour palette.}
\item{full}{The full set of all six colours in the \acronym{AAGI}
palette}
\item{colourful}{All colours in the palette minus \dQuote{AAGI Black}}
\item{bright}{Four \dQuote{bright} colours including,
\dQuote{AAGI Bright Green}, \dQuote{AAGI Blue} and \dQuote{AAGI Yellow}}
\item{bluegreeen}{Three blue, green or bluegreen \acronym{AAGI} colours
including \dQuote{AAGI Bright Green}, \dQuote{AAGI Turquoise},
\dQuote{AAGI Blue}, \dQuote{AAGI Green} and \dQuote{AAGI Bluegreen}}
}
}
\examples{
# View all five palettes' names and hexadecimal colour codes
 aagi_palettes()
 aagi_palettes()$colourful
 aagi_palettes()$bluegreen

}
