% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{aagipal}
\alias{aagipal}
\title{Interpolate an AAGI Colourway or Palette}
\usage{
aagipal(colourway = NULL, reverse = FALSE, ...)
}
\arguments{
\item{colourway}{\code{Character} name of a \acronym{AAGI} colourway in
\code{aagicolourways()}.  Defaults to \code{NULL} returning a randomly selected
colourway.}

\item{reverse}{\code{Boolean} argument indicating whether the colourway should be
reversed.  Defaults to \code{FALSE}.}

\item{...}{Additional arguments to pass to \code{colorRampPalette()}.}
}
\description{
Interpolate an AAGI Colourway or Palette
}
