% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aagi_logo.R
\name{add_aagi_logo}
\alias{add_aagi_logo}
\title{Insert AAGI's Logo Into Any Graphic File}
\usage{
add_aagi_logo(file_in, file_out, logo_width = 4.6, overwrite = FALSE)
}
\arguments{
\item{file_in}{Path to the graphical file to import and add the
\acronym{AAGI} logo to apply the logo to.}

\item{file_out}{File name to create on disk as a .png format image.}

\item{logo_width}{Size for the logo in centimetres.  Defaults to 4.6 cm, the
smallest allowed by the \acronym{AAGI} guidelines. Larger sizes may be
beneficial for larger graphical images than normal reporting and sharing.}

\item{overwrite}{A Boolean value that indicates whether to overwrite an
existing file or not. Defaults to \code{FALSE} and will not overwrite the
existing file.}
}
\value{
An invisible \code{NULL}, called for its side effect of adding the
\acronym{AAGI} logo to the plot that's called.
}
\description{
Insert the \acronym{AAGI} logo in any graphical image.  Particularly well-
suited for use with files created with \pkg{AAGIThemes}.  The logo will be
inserted in the upper-left of the image with a size of at least 4.6 cm as
per \acronym{AAGI} brand guidelines.
}
\section{Image Sizing}{
 This function checks the image's DPI values and sets
the logo size accordingly. When saving, e.g., from \CRANpkg{ggplot}, specify
the DPI value along with your height and width and units for optimal
performance.
`ggsave(
filename = "AAGI.png",
plot = p1,
path = tempdir(),
width = 18,
height = 18,
units = "cm",
dpi = 300
)
}

\examples{
library("ggplot2")

p1 <- ggplot(mtcars) +
  geom_point(aes(
    x = wt,
    y = mpg,
    colour = factor(gear)
  )) +
  facet_wrap(~am) +
  theme_aagi()

ggsave(p1, filename = "AAGI.png", path = tempdir())

add_aagi_logo(
  file_in = file.path(tempdir(), "AAGI.png"),
  file_out = file.path(tempdir(), "AAGI_logo.png")
)

x <- magick::image_read(file.path(tempdir(), "AAGI_logo.png"))
print(x)

}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
