% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdm_colours.R
\docType{data}
\name{ccdm_colours}
\alias{ccdm_colours}
\title{CCDM Colours}
\format{
A \code{vector} with six (6) colours as hexadecimal values:
\describe{
\item{CCDM Blue}{#648FD2}
\item{CCDM Green}{#54921E}
\item{CCDM Bluegreen}{#31859C}
\item{CCDM Yellow}{#FFBC42}
\item{CCDM Light Grey}{#A6A6A6}
\item{CCDM Dark Grey}{#414042}
}
}
\usage{
ccdm_colours
}
\description{
A \code{vector} containing six official \acronym{CDDM} colours as hexadecimal
codes for use in plotting.
}
\keyword{datasets}
