% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_as_hex.R
\name{colour_as_hex}
\alias{colour_as_hex}
\title{Extract AAGI Colours as Hex Codes From Names}
\usage{
colour_as_hex(...)
}
\arguments{
\item{...}{Character names of \acronym{AAGI} colours}
}
\description{
Extract AAGI Colours as Hex Codes From Names
}
