% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_aagi_palette.R
\name{display_aagi_palette}
\alias{display_aagi_palette}
\title{AAGI Palettes}
\usage{
display_aagi_palette(n = 3, name)
}
\arguments{
\item{n}{\code{Numeric} Number of different colours in the palette, minimum 3,
maximum depends on the palette. Defaults to \code{3}}

\item{name}{A palette name from the lists below

There are 3 types of palettes, sequential, diverging, and qualitative.
\enumerate{
\item{Sequential palettes are suited to ordered data that progress from low
to high. Lightness steps dominate the look of these schemes, with light
colours for low data values to dark colours for high data values.}
\item{Diverging palettes put equal emphasis on mid-range critical values and
extremes at both ends of the data range. The critical class or break in the
middle of the legend is emphasized with light colours and low and high
extremes are emphasized with dark colours that have contrasting hues.}
\item{Qualitative palettes do not imply magnitude differences between legend
classes, and hues are used to create the primary visual differences between
classes. Qualitative schemes are best suited to representing nominal or
categorical data.}
}

The sequential palettes names are:
\itemize{
\item \dQuote{aagi_blues},
\item \dQuote{aagi_bright_greens},
\item \dQuote{aagi_BrYl},
\item \dQuote{aagi_greens},
\item \dQuote{aagi_greys},
\item \dQuote{aagi_reds},
\item \dQuote{aagi_oranges},
\item \dQuote{aagi_RdYl},
\item \dQuote{aagi_teals},
\item \dQuote{aagi_TlGnYl},
\item \dQuote{aagi_TlYl},
\item \dQuote{aagi_yellows}
}

All the sequential palettes are available in variations from 3 different
values up to 9 different values.

The diverging palettes are:
\itemize{
\item \dQuote{aagi_BuOr},
\item \dQuote{aagi_BuYl},
\item \dQuote{aagi_GnYl},
\item \dQuote{aagi_RdBu},
\item \dQuote{aagi_RdTl},
\item \dQuote{aagi_TlGn}
}

All the diverging palettes are available in variations from 3 different
values up to 11 different values.

For qualitative palettes, the lowest number of distinct values available
always is 3, but the largest number is different for different palettes.
It is given together with the palette names in the following table.}
}
\value{
\code{NULL}. Called for its side-effects, a plot of the given palette.
}
\description{
A quick and dirty way to show \acronym{AAGI} colours in a paelette plot.
}
\examples{
display_aagi_pal(n = 5, name = "aagi_BuOr")

}
