% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{..set_aagi_font}
\alias{..set_aagi_font}
\title{Set AAGI Font for Figures and Tables With Fallback}
\usage{
..set_aagi_font()
}
\value{
A character string with a font family name.
}
\description{
Sets the AAGI font to Proxima Nova first; if it's not available, falls back
to Arial as per the \acronym{AAGI} comms guide. If neither is available,
returns "sans".
}
\keyword{internal}
