% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_aagi.R
\name{.aagi_colour_order}
\alias{.aagi_colour_order}
\title{Set a different default order of AAGI colours for better data vis}
\usage{
.aagi_colour_order()
}
\description{
Sets the colour order so that there is more contrast between the first three
colours since graphs frequently only have 1-3 values displayed.
}
\keyword{internal}
