% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.choose_aagi_font}
\alias{.choose_aagi_font}
\title{Choose the first available font from a preferred list}
\usage{
.choose_aagi_font(preferred = c("Proxima Nova", "Arial"))
}
\arguments{
\item{preferred}{Character vector of font family names.}
}
\value{
Font family name, or NA_character_ if none are available.
}
\description{
Choose the first available font from a preferred list
}
\keyword{internal}
