% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_aagi_colour}
\alias{.convert_aagi_colour}
\title{Check and convert AAGI colour}
\usage{
.convert_aagi_colour(x)
}
\arguments{
\item{x}{A character string to be checked/converted. Can be length-1 or a
vector.}
}
\value{
A character string/vector of hex colour codes, or the input
unchanged if it's not an AAGI colour name.
}
\description{
Checks a string value to see if it's a valid \acronym{AAGI} colour and
converts it to the corresponding hexadecimal colour code. This is done by
checking for an "AAGI " prefix in the string, as all valid colours from
\pkg{AAGIPalettes} will have this.
}
\keyword{internal}
