% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_aagi.R
\name{.normalise_hist_breaks}
\alias{.normalise_hist_breaks}
\title{Normalise the \code{breaks} argument for hist_aagi()}
\usage{
.normalise_hist_breaks(breaks)
}
\arguments{
\item{breaks}{User-supplied breaks selector.}
}
\value{
One of "pretty", "scott", or "exact".
}
\description{
Normalise the \code{breaks} argument for hist_aagi()
}
\keyword{internal}
