% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.register_font}
\alias{.register_font}
\title{Register a font with sysfonts (and Windows device, if applicable)}
\usage{
.register_font(font_name)
}
\arguments{
\item{font_name}{Font family name to register.}
}
\value{
TRUE if registered, FALSE otherwise.
}
\description{
Register a font with sysfonts (and Windows device, if applicable)
}
\keyword{internal}
