% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_aagi_pal.R
\name{interpolate_aagi_pal}
\alias{interpolate_aagi_pal}
\title{Interpolate an AAGI Colour Palette}
\usage{
interpolate_aagi_pal(
  aagi_palette = c("colourful", "primary", "secondary", "sequential", "diverging"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{aagi_palette}{(character; default = \code{graph}) given name of an
\acronym{AAGI} palette: \code{\link[=make_aagi_palettes]{make_aagi_palettes()}}}

\item{reverse}{(Boolean; default = \code{FALSE}) indicating if palette should
be reverse}

\item{...}{Additional arguments to pass to \code{colorRampPalette} see
details here \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}.}
}
\value{
A function that takes a single value and makes that many colours.
}
\description{
This function takes a \acronym{AAGI} colour palette and generates more
colours from it, so that there are enough to make your chart.
}
\details{
The interpolation method is set to \code{spline} (the default is \code{linear}) in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = as.character(wt))) +
  geom_point() +
  scale_colour_manual(values = make_aagi_palette()(29))

}
\seealso{
\code{\link{grattan_palettes}}
}
