% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/aagi_pal.R, R/make_aagi_palette.R
=======
% Please edit documentation in R/make_aagi_palette.R
>>>>>>> dfa1032 (Fixes broken shit)
\name{make_aagi_palette}
\alias{make_aagi_palette}
\title{Interpolate an AAGI Colour Palette}
\usage{
make_aagi_palette(
<<<<<<< HEAD
  make_aagi_palette = c("colourful", "primary", "secondary", "sequential", "diverging"),
  reverse = FALSE,
  ...
)

make_aagi_palette(
  make_aagi_palette = c("colourful", "primary", "secondary", "sequential", "diverging"),
=======
  make_aagi_paletteette = c("colourful", "primary", "secondary", "sequential",
    "diverging"),
>>>>>>> dfa1032 (Fixes broken shit)
  reverse = FALSE,
  ...
)
}
\arguments{
\item{reverse}{(\code{Boolean}; default = \code{FALSE}) indicating if palette should
be reverse}

\item{...}{Additional arguments to pass to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}},
see details here \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}.}

\item{palette}{(\code{character}; default = \code{graph}) given name of an
\acronym{AAGI} palette: \code{\link[=make_aagi_palettes]{make_aagi_palettes()}}}
}
\value{
A function that takes a single value and makes that many colours.
<<<<<<< HEAD

A function that takes a single value and makes that many colours.
=======
>>>>>>> dfa1032 (Fixes broken shit)
}
\description{
This function takes a \acronym{AAGI} colour palette and generates more
colours from it, so that there are enough to make your chart.
<<<<<<< HEAD

This function takes a \acronym{AAGI} colour palette and generates more
colours from it, so that there are enough to make your chart.
=======
>>>>>>> dfa1032 (Fixes broken shit)
}
\details{
The interpolation method is set to \code{spline} (the default is \code{linear}) in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.
<<<<<<< HEAD

The interpolation method is set to \code{spline} (the default is \code{linear}) in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.
=======
>>>>>>> dfa1032 (Fixes broken shit)
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = as.character(wt))) +
  geom_point() +
<<<<<<< HEAD
  scale_colour_manual(values = make_aagi_palette()(29))

library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = as.character(wt))) +
  geom_point() +
  scale_colour_manual(values = make_aagi_palette()(29))

}
\seealso{
\code{\link{grattan_palettes}}

=======
  scale_colour_manual(values = make_make_aagi_palette()(29))

}
\seealso{
>>>>>>> dfa1032 (Fixes broken shit)
aagi_palettes
}
