% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_aagi.R
\name{pal_aagi}
\alias{pal_aagi}
\alias{aagi_pal}
\title{Colour AAGI Palette (Discrete)}
\usage{
pal_aagi(type = "seq", palette = 1, direction = 1)

aagi_pal(type = "seq", palette = 1, direction = 1)
}
\arguments{
\item{type}{One of \dQuote{seq} (sequential), \dQuote{div} (diverging) or
\dQuote{qual} (qualitative).}

\item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}.}

\item{direction}{Sets the order of colours in the scale. If 1, the default,
colours are as output by \code{\link[=aagi_pal]{aagi_pal()}}. If \code{-1}, the order of colours is
reversed.}
}
\description{
Colour AAGI Palette (Discrete)
}
\examples{
show_col(pal_aagi()(10))
show_col(pal_aagi("div")(5))
show_col(pal_aagi(palette = "Greens")(5))

# Can use with gradient_n to create a continuous gradient
cols <- pal_aagi("div")(5)
show_col(pal_gradient_n(cols)(seq(0, 1, length.out = 30)))
}
