% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par_aagi}
\alias{par_aagi}
\title{Set Graphical Parameters That Satisfy AAGI's Style Requirements}
\usage{
par_aagi()
}
\value{
A list object of graphical parameters
}
\description{
Set Graphical Parameters That Satisfy AAGI's Style Requirements
}
\keyword{internal}
