% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aagi.R
\name{showtext_aagi_off}
\alias{showtext_aagi_off}
\title{Disable global showtext auto rendering (AAGIThemes)}
\usage{
showtext_aagi_off()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
\code{theme_aagi(showtext_auto = TRUE)} enables \code{showtext::showtext_auto()}, which
affects subsequent plots globally. Call this function to turn it off.
}
