% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.aagi-package.R
\docType{package}
\name{theme.aagi-package}
\alias{theme.aagi}
\alias{theme.aagi-package}
\title{theme.aagi: Apply AAGI Brand Guidelines to Graphic Output and Create AAGI Themed Documents}
\description{
Applies 'AAGI' external brand guidelines to graphics. 'AAGI' colours and font guidelines are applied as useful and reasonable to both base graphics as well as 'ggplot2' figures. Support for creating MS Word and PowerPoint files using Quarto and RMarkdown templates is also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCDM-AAGI/theme.aagi}
  \item Report bugs at \url{https://github.com/CCDM-AAGI/theme.aagi/issues}
}

}
\author{
\strong{Maintainer}: Tinula Kariyawasam \email{tinula.kariyawasam@dpird.wa.gov.au}

Authors:
\itemize{
  \item Adam H. Sparks \email{adam.sparks@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID}) [conceptor]
  \item Kenyon Ng \email{kenyon.ng@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-6315-9831}{ORCID})
  \item Matt Cowgill \email{matthew.cowgill@grattaninstitute.edu.au} (Wrote major portions of package for saving charts
 for 'Grattan' package, adapted to 'theme.aagi' under 'MIT'
 licence.)
  \item Will Mackey \email{william.mackey@grattaninstitute.edu.au} (Wrote major portions of package for saving charts
 for 'Grattan' package, adapted to 'theme.aagi' under 'MIT'
 licence.)
}

Other contributors:
\itemize{
  \item Jimmy Ng \email{jimmy.ng@dpird.wa.gov.au} [contributor]
  \item Curtin University of Technology [copyright holder]
}

}
\keyword{internal}
