% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Add a Watermark Annotation Layer to a ggplot2 Object}
\usage{
watermark(
  watermark,
  fontsize = 120,
  colour = "grey90",
  alpha = 0.1,
  fontface = "bold",
  angle = 22
)
}
\arguments{
\item{watermark}{String to be added as watermark}

\item{fontsize}{Font size}

\item{colour}{Font colour}

\item{alpha}{Alpha (transparency; lower number = more transparent)}

\item{fontface}{Font face ("bold" by default)}

\item{angle}{Angle of the watermark}
}
\value{
A \link[ggplot2:ggplot2]{ggplot2} object with a watermark added to the plot that's called.
}
\description{
Add a Watermark Annotation Layer to a ggplot2 Object
}
\examples{

library(ggplot2)
# First, start with a plot:
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_aagi()

# Then add a watermark:

p + watermark("DRAFT")

}
\author{
Matt Cowgill and Will Mackey

Adam H. Sparks, \email{adam.sparks@curtin.edu.au}
}
