% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_colours.R, R/data.R
\docType{data}
\name{aagi_colours}
\alias{aagi_colours}
\title{AAGI Colours}
\format{
A \code{vector} with three (3) colours as hexadecimal values:
\describe{
\item{AAGI Black}{#414042}
\item{AAGI Bright Green}{#B6D438}
\item{AAGI Turquoise}{#00808B}
\item{AAGI Blue}{#648FD2}
\item{AAGI Green}{#54921E}
\item{AAGI Yellow}{#FFBC42}
}

A \link{vector} object of 6 items.
}
\usage{
aagi_colours

aagi_colours
}
\description{
A \code{vector} containing three official \acronym{AAGI} colours as hexadecimal
codes for use in plotting.

A \link{vector} object containing six (6) items representing AAGI colours as
hexadecimal values.
}
\concept{data}
\keyword{datasets}
