% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_colours.R
\docType{data}
\name{aagi_colours}
\alias{aagi_colours}
\title{AAGI Colours}
\format{
A \code{vector} with three (3) colours as hexadecimal values:
\describe{
\item{AAGI Black}{#414042}
\item{AAGI Bright Green}{#B6D438}
\item{AAGI Turquoise}{#00808b}
}
}
\usage{
aagi_colours
}
\description{
A \code{vector} containing three official \acronym{AAGI} colours as hexadecimal
codes for use in plotting.
}
\keyword{datasets}
