% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_cols.R
\name{aagi_cols}
\alias{aagi_cols}
\title{Extract AAGI Colours as Hex Codes}
\usage{
aagi_cols(...)
}
\arguments{
\item{...}{Character names of \acronym{AAGI} colours}
}
\description{
Extract AAGI Colours as Hex Codes
}
