% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{aagi_pal}
\alias{aagi_pal}
\title{Interpolate an AAGI Palette}
\usage{
aagi_pal(palette = NULL, reverse = FALSE, ...)
}
\arguments{
\item{reverse}{\code{Boolean} argument indicating whether the palette should be
reversed, \emph{i.e.} light to dark.  Defaults to \code{FALSE}.}

\item{...}{Additional arguments to pass to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}

\item{Palette}{\code{Character} name of a \acronym{AAGI} palette in
\code{aagi_palettes()}.  Defaults to \code{NULL} returning a randomly selected
palette.}
}
\description{
Interpolate an \acronym{AAGI} palette to have more colours for use in
graphing.  This should be used sensibly and within reason, \emph{e.g.}, no 15
colour graphs, 6 or 7 unique colours is a sensible value for discrete
variables.  Luckily the \link{aagi_colours} have 6!  However, with this
function, you can create new palettes by interpolating the colours in
\code{aagi_colours}. Also see \CRANpkg{RColorBrewer} and \CRANpkg{viridis} for
sensible palettes.
}
