% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.aagi-package.R
\docType{package}
\name{AAGIThemes-package}
\alias{AAGIThemes}
\alias{AAGIThemes-package}
\title{AAGIThemes: AAGI Branding for Graphical and Tabular Outputs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Applies 'AAGI' external brand guidelines to graphics. 'AAGI' colours and font guidelines are applied as useful and reasonable to base graphics, 'ggplot2' figures and 'flextable' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AAGI-AUS/AAGIThemes}
  \item \url{https://AAGI-AUS.github.io/AAGIThemes/}
  \item Report bugs at \url{https://github.com/AAGI-AUS/AAGIThemes/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adam.sparks@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Kenyon Ng \email{kenyon.ng@dpird.wa.gov.au} (\href{https://orcid.org/0000-0002-6315-9831}{ORCID})
  \item Tinula Kariyawasam \email{tinula.kariyawasam@dpird.wa.gov.au}
  \item Matt Cowgill \email{matthew.cowgill@grattaninstitute.edu.au} (Wrote major portions of package for saving charts
 for 'Grattan' package, adapted to 'AAGIThemes' under 'MIT'
 licence.)
  \item Will Mackey \email{william.mackey@grattaninstitute.edu.au} (Wrote major portions of package for saving charts
 for 'Grattan' package, adapted to 'AAGIThemes' under 'MIT'
 licence.)
}

Other contributors:
\itemize{
  \item Jimmy Ng \email{jimmy.ng@dpird.wa.gov.au} [contributor]
  \item Curtin University of Technology (Provided support through Adam Sparks's time.) [copyright holder]
  \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
