% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aagi_logo.R
\name{add_aagi_logo}
\alias{add_aagi_logo}
\title{Insert AAGI's Logo Into Any Graphic File}
\usage{
add_aagi_logo(file_in, file_out, logo_scale = 2.5)
}
\arguments{
\item{file_in}{Path to the graphical file to import and add the AAGI logo
to.}

\item{file_out}{File name to create on disk as a .png format image.}

\item{logo_scale}{Scaling for the logo.  Defaults to 2.5, larger sizes may be
beneficial for larger graphical images than normal reporting and sharing.
Care should be taken that the logo follows the CCDM branding guidelines
and the size remains at least 4.6 cm.}
}
\description{
Insert the \acronym{AAGI} logo in any graphical image.  Particularly well-
suited for use with files created with \pkg{theme.aagi}.  The logo will be
inserted in the upper-left of the image with a size of at least 4.6 cm as
per \acronym{AAGI} brand guidelines.
}
\examples{
library("ggplot2")

p1 <- ggplot(mtcars) +
  geom_point(aes(
    x = wt,
    y = mpg,
    colour = factor(gear)
   )) +
  facet_wrap(~ am) +
  theme_aagi()

ggsave(p1, filename = "AAGI.png", path = tempdir())

add_aagi_logo(file_in = file.path(tempdir(), "AAGI"),
               file_out = file.path(tempdir(), "AAGI_logo.png"))

x <- magick::image_read(file.path(tempdir(), "AAGI_logo.png"))
print(x)

}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
