% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_aagi.R
\name{barplot_aagi}
\alias{barplot_aagi}
\title{Basic Barplots Using a Unified AAGI Style and Typography}
\usage{
barplot_aagi(height, ...)
}
\arguments{
\item{height}{Either a vector or matrix of values describing the bars which
make up the plot.  If height is a vector, the plot consists of a sequence
of rectangular bars with heights given by the values in the vector.  If
height is a matrix and beside is \code{FALSE} then each bar of the plot
corresponds to a column of height, with the values in the column giving the
heights of stacked sub-bars making up the bar.  If height is a matrix and
beside is \code{TRUE}, then the values in each column are juxtaposed rather than
stacked.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).  The most commonly used argument would be
\code{y} for the factor to use for the y-axis, \code{type} describing what type of
plot should be drawn, or \code{col} for point colour (defaults to AAGI Black),
a very dark grey.}
}
\value{
A \code{barplot} object, returned invisibly (see \code{\link[graphics:barplot]{graphics::barplot()}}).
}
\description{
Basic barplots that follow a standard \acronym{AAGI} style
including typography guidelines that uses (hopefully) sensible defaults.
All valid \code{barplot()} options are supported through \code{...}, for \emph{e.g.},
\code{col} to set the colour.  Defaults to "AAGI Black", a very dark grey
colour.
}
\examples{

barplot_aagi(islands)
barplot_aagi(islands, col = "AAGI Orange")

}
\seealso{
\itemize{
\item \code{\link[graphics:plot]{graphics::plot()}} for full documentation of the basic plotting
capabilities.
\item plot_aagi, boxplot_aagi, hist_aagi
}

Other Baseplots:
\code{\link[=boxplot_aagi]{boxplot_aagi()}},
\code{\link[=hist_aagi]{hist_aagi()}},
\code{\link[=plot_aagi]{plot_aagi()}}
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
\concept{Baseplots}
