% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_aagi.R
\name{boxplot_aagi}
\alias{boxplot_aagi}
\title{Basic Boxplots Using a Unified AAGI Style and Typography}
\usage{
boxplot_aagi(x, main = "", sub = "", xlab = "", ylab = "", pch = 16, ...)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced.
Either a numeric vector, or a single list containing such vectors.
Additional unnamed arguments specify further data as separate vectors (each
corresponding to a component boxplot).  \code{NA}s are allowed in the data.}

\item{main}{Main title.  Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label.  Optional, if not supplied it will
be blank.}

\item{xlab}{X-axis label.  Optional.}

\item{ylab}{Y-axis label.  Optional.}

\item{pch}{plotting 'character', \emph{i.e.}, symbol to use.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).  The most commonly used argument would be
\code{y} for the factor to use for the y-axis, \code{type} describing what type of
plot should be drawn, or \code{col} for point colour (defaults to AAGI Black),
a very dark grey.}

\item{col}{Colour to use as fill for boxes  Defaults to white.  Can be
supplied as a named AAGI colour, \emph{e.g.}, "AAGI Orange"; named colour,
"Orange"; or or a hexadecimal code, \emph{e.g.}, "#ec8525".}
}
\value{
Called for its side effect of creating a boxplot with the
\acronym{AAGI} style.
}
\description{
Basic boxplots that follow \acronym{AAGI} colour and typography guidelines
that uses (hopefully) sensible defaults.
}
\examples{
boxplot_aagi(decrease ~ treatment,
  data = OrchardSprays,
  xlab = "treatment",
  ylab = "decrease"
)

}
\seealso{
\itemize{
\item \code{\link[graphics:boxplot]{graphics::boxplot()}} for full documentation of the basic boxplot
capabilities, and
\item \code{\link[graphics:par]{graphics::par()}} for full documentation of \code{pch}.
}

\itemize{
\item \code{\link[graphics:boxplot]{graphics::boxplot()}} for full documentation of the basic plotting
capabilities.
\item barplot_aagi, hist_aagi, plot_aagi
}

Other Baseplots:
\code{\link[=barplot_aagi]{barplot_aagi()}},
\code{\link[=hist_aagi]{hist_aagi()}},
\code{\link[=plot_aagi]{plot_aagi()}}
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
\concept{Baseplots}
