% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_aagi.R
\name{boxplot_aagi}
\alias{boxplot_aagi}
\title{Basic Boxplots Using a Unified AAGI Style and Typography}
\usage{
boxplot_aagi(
  x,
  main = "",
  sub = "",
  xlab = "",
  ylab = "",
  col = "white",
  pch = 16,
  ...
)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced.
Either a numeric vector, or a single list containing such vectors.
Additional unnamed arguments specify further data as separate vectors (each
corresponding to a component boxplot).  \code{NA}s are allowed in the data.}

\item{main}{Main title.  Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label.  Optional, if not supplied it will
be blank.}

\item{xlab}{X-axis label.  Optional.}

\item{ylab}{Y-axis label.  Optional.}

\item{col}{Colour to use as fill for boxes  Defaults to white.  Can be
supplied as a named AAGI colour, \emph{e.g.}, "AAGI Orange"; named colour,
"Orange"; or or a hexadecimal code, "#ec8525".}

\item{pch}{plotting ‘character’, \emph{i.e.}, symbol to use. This can either
be a single character or an integer code for one of a set of graphics
symbols. The full set of S symbols is available with \code{pch = 0:18}, see the
examples below. (NB: \R uses circles instead of the octagons used in S.)
Value \code{pch = "."} (equivalently \code{pch = 46}) is handled specially. It is a
rectangle of side 0.01 inch (scaled by \code{cex}). In addition, if \code{cex = 1}
(the default), each side is at least one pixel (1/72 inch on the
\code{\link[grDevices:pdf]{grDevices::pdf()}}, \code{\link[grDevices:postscript]{grDevices::postscript()}} and \code{\link[grDevices:xfig]{grDevices::xfig()}}
devices).
For other text symbols, \code{cex = 1} corresponds to the default font size of
the device, often specified by an argument \code{pointsize}. For \code{pch} in 0:25
the default size is about 75\% of the character height (see \code{par("cin")}).}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).}
}
\value{
Called for its side effect of creating a boxplot with the
\acronym{AAGI} style.
}
\description{
Basic boxplots that follow \acronym{AAGI} colour and typography guidelines
that uses (hopefully) sensible defaults.
}
\examples{
boxplot_aagi(decrease ~ treatment,
  data = OrchardSprays,
  xlab = "treatment",
  ylab = "decrease"
)
}
\seealso{
\code{\link[graphics:boxplot]{graphics::boxplot()}} for full documentation of the basic boxplot
capabilities
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
