% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convert_aagi_colour}
\alias{.convert_aagi_colour}
\title{Check and convert AAGI colour}
\usage{
.convert_aagi_colour(x)
}
\arguments{
\item{x}{A character string name value to be checked if it's an AAGI colour
and converted to the corresponding hex value. Can be a single value or a
vector.}
}
\value{
A character string or vector of hexadecimal colour codes, or the
input unchanged if it's not an AAGI colour name.
}
\description{
Check and convert AAGI colour
}
\keyword{internal}
