% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_aagi.R
\name{hist_aagi}
\alias{hist_aagi}
\title{Basic Histograms Using a Unified AAGI Style and Typography}
\usage{
hist_aagi(
  x,
  main = "",
  sub = "",
  xlab = "",
  ylab = "Count",
  breaks = "pretty",
  ...
)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{main}{Main title.  Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label.  Optional, if not supplied it will
be blank.}

\item{xlab}{X-axis label.  Optional, if not supplied will be an empty string
with no value.}

\item{ylab}{Y-axis label.  Optional, if not supplied will default to "Count".}

\item{breaks}{Either \code{pretty}, default, or \code{exact}.  Pretty uses Scott's
Rule, whereas exact uses a bin-width of 1 so each value is represented
individually with a bar.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).  The most commonly used argument would be
\code{y} for the factor to use for the y-axis, \code{type} describing what type of
plot should be drawn, or \code{col} for point colour (defaults to AAGI Black),
a very dark grey.}
}
\value{
\code{NULL}, returned invisibly, consistent with \code{\link[graphics:hist]{graphics::hist()}}.
Called primarily for its side effect of creating a plot with the
\acronym{AAGI} style.
}
\description{
Basic histograms that follow a standard \acronym{AAGI} style
including typography guidelines that uses (hopefully) sensible defaults.
All valid \code{boxplot()} options are supported through \code{...}, for \emph{e.g.},
\code{col} to set the colour.  Defaults to "AAGI Black", a very dark grey
colour.
}
\examples{

hist_aagi(islands)

}
\references{
Scott, D.W. (1979) On optimal and data-based histograms. \emph{Biometrika},
66, 605–610.

Hyndman, R.J., The Problem with Sturges' Rule for Constructing Histograms,
1995, Monash University,
\url{https://www.robjhyndman.com/papers/sturges.pdf}.
}
\seealso{
\itemize{
\item \code{\link[graphics:hist]{graphics::hist()}} for full documentation of the basic histogram
capabilities.
\item barplot_aagi, boxplot_aagi, plot_aagi
}

Other Baseplots:
\code{\link[=barplot_aagi]{barplot_aagi()}},
\code{\link[=boxplot_aagi]{boxplot_aagi()}},
\code{\link[=plot_aagi]{plot_aagi()}}
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
\concept{Baseplots}
