% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_aagi.R
\name{hist_aagi}
\alias{hist_aagi}
\title{Basic Histograms Using a Unified DPIRD Style and Typography}
\usage{
hist_aagi(
  x,
  main = "",
  sub = "",
  xlab = "",
  ylab = "Count",
  breaks = "pretty",
  col = "#414042",
  ...
)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{main}{Main title.  Optional, if not supplied it will be blank.}

\item{sub}{Sub title below x-axis label.  Optional, if not supplied it will
be blank.}

\item{xlab}{X-axis label.  Optional, if not supplied will be taken from data.}

\item{ylab}{Y-axis label.  Optional, if not supplied will default to "Count".}

\item{breaks}{Either \code{pretty}, default, or \code{exact}.  Pretty uses Scott's
Rule, whereas exact uses a bin-width of 1 so each value is represented
individually with a bar.}

\item{col}{Colour to use as fill for bars  Defaults to a very dark grey.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).}
}
\description{
Basic histograms that follow a standard \acronym{AAGI} style
including typography guidelines that uses (hopefully) sensible defaults.
}
\examples{

hist_aagi(islands)

}
\references{
Scott, D.W. (1979) On optimal and data-based histograms. \emph{Biometrika},
66, 605–610.

Hyndman, R.J., The Problem with Sturges' Rule for Constructing Histograms,
1995, Monash University,
\url{https://www.robjhyndman.com/papers/sturges.pdf}.
}
\seealso{
\code{\link[graphics:hist]{graphics::hist()}} for full documentation of the basic histogram
capabilities
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
