% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aagi_palette.R
\name{make_aagi_palette}
\alias{make_aagi_palette}
\title{Interpolate an AAGI Colour Palette}
\usage{
make_aagi_palette(
  palette = c("colourful", "primary", "secondary", "sequential", "diverging"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{(\code{character}; default = \code{colourful}) given name of an
\acronym{AAGI} palette: \code{\link[=aagi_palettes]{aagi_palettes()}}}

\item{reverse}{(\code{Boolean}; default = \code{FALSE}) indicating if palette should
be reversed.}

\item{...}{Additional arguments to pass to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}},
see details here \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}.}
}
\value{
A function that takes a single value and makes that many colours.
}
\description{
This function takes a \acronym{AAGI} colour palette and generates more
colours from it, so that there are enough to make your chart. The
interpolation method is set to \code{spline} (the default is \code{linear}) in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp, colour = as.character(wt))) +
  geom_point() +
  scale_colour_manual(values = make_aagi_palette())

}
\seealso{
aagi_palettes
}
