% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aagi.R
\name{plot_aagi}
\alias{plot_aagi}
\title{Basic X-Y Plotting Using a Unified AAGI Style and Typography}
\usage{
plot_aagi(x, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a single
plotting structure, function or any \R \emph{object with a plot method} can be
provided.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \link[graphics:par]{graphics::par}).  The most commonly used argument would be
\code{y} for the factor to use for the y-axis or \code{type} describing what type of
plot should be drawn.}
}
\value{
Called for its side effect of creating a plot with the
\acronym{AAGI} style.
}
\description{
Basic X-Y plotting that follows \acronym{AAGI} colour and
typography guidelines that uses (hopefully) sensible defaults.
}
\examples{
library(datasets)

plot_aagi(airquality$Ozone,
  main = "Air Quality",
  xlab = "Ozone"
)

plot_aagi(pressure)

}
\seealso{
\link[graphics:plot.default]{graphics::plot} for full documentation of the basic plotting
capabilities.
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
