% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aagi.R
\name{plot_aagi}
\alias{plot_aagi}
\title{Basic X-Y Plotting Using a Unified AAGI Style and Typography}
\usage{
plot_aagi(
  x,
  y = NULL,
  main = "",
  sub = "",
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  col = AAGIPallets::aagi_colours["AAGI Black"],
  pch = 16,
  type = "p",
  ...
)
}
\arguments{
\item{x}{Values to be plotted on x-axis.}

\item{y}{Values to be used on y-axis.  Optional.}

\item{main}{Main title text.}

\item{sub}{Subtitle text below x-axis label.}

\item{xlab}{X-axis label.  Optional, if blank will default to data supplied
label.}

\item{ylab}{Y-axis label.  Optional, if blank will default to data supplied
label if one is available.}

\item{xlim}{Numeric value for x-axis limits.  Optional.}

\item{ylim}{Numeric value for y-axis limits.  Optional.}

\item{col}{Colour to use for points.  Defaults to "AAGI Black", a very dark
grey.}

\item{pch}{plotting ‘character’, \emph{i.e.}, symbol to use. This can either
be a single character or an integer code for one of a set of graphics
symbols. The full set of S symbols is available with \code{pch = 0:18}, see the
examples below. (NB: \R uses circles instead of the octagons used in S.)}

\item{type}{What type of plot should be drawn.  Defaults to \code{p} for
\strong{p}oints``  See Section "Type" for more.}

\item{...}{Arguments to be passed to methods, such as graphical parameters
(see \code{\link[graphics:par]{graphics::par()}}).  The most commonly used argument would be
\code{type} what type of plot should be drawn.}
}
\description{
Basic X-Y plotting that follows \acronym{AAGI} colour and
typography guidelines that uses (hopefully) sensible defaults.
}
\section{Type}{

Possible types are:
\describe{
\item{\dQuote{p}}{for \strong{p}oints}
\item{\dQuote{l}}{for \strong{l}ines}
\item{\dQuote{b}}{for \strong{b}oth}
\item{\dQuote{c}}{for the lines part alone of \dQuote{\strong{b}}}
\item{\dQuote{o}}{for both \sQuote{\strong{o}verplotted}}
\item{\dQuote{h}}{for \sQuote{\strong{h}istogram} like (or
\sQuote{\strong{h}igh-density}) vertical lines}
\item{\dQuote{s}}{for stair \strong{s}teps}
\item{\dQuote{n}}{for \strong{n}o plotting}
}

Value \code{pch = "."} (equivalently \code{pch = 46}) is handled specially. It is a
rectangle of side 0.01 inch (scaled by \code{cex}). In addition, if \code{cex = 1}
(the default), each side is at least one pixel (1/72 inch on the
\code{\link[grDevices:pdf]{grDevices::pdf()}}, \code{\link[grDevices:postscript]{grDevices::postscript()}} and \code{\link[grDevices:xfig]{grDevices::xfig()}}
devices).

For other text symbols, \code{cex = 1} corresponds to the default fontsize of the
device, often specified by an argument \code{pointsize}. For \code{pch} in 0:25 the
default size is about 75\% of the character height (see \code{par("cin")}).
}

\examples{
library(datasets)

plot_aagi(airquality$Ozone,
          main = "Air Quality",
          xlab = "Ozone")

}
\seealso{
\code{\link[graphics:plot.default]{graphics::plot()}} for full documentation of the basic plotting capabilities.
}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
