% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_ggplot2_scales.R
\name{scale_colour_aagi}
\alias{scale_colour_aagi}
\title{Construct AAGI Colour Scales for ggplot2}
\usage{
scale_colour_aagi(palette = "colourful", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{\code{Character} string, one of:
\itemize{
\item "primary",
\item "secondary",
\item "colourful" (default),
\item "full",
\item "bright", or
\item "bluegreen".
}}

\item{discrete}{\code{Boolean} argument indicating whether colour aesthetic is
discrete, \emph{e.g.}, a \code{factor} (\code{TRUE}) or continuous, \emph{e.g.}, \code{numeric}
(\code{FALSE}).}

\item{reverse}{\code{Boolean} argument indicating whether the palette should be
reversed, \emph{i.e.}, light to dark.  Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{discrete_scale()} or
\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn}, used respectively when
discrete is \code{TRUE}.}
}
\description{
Uses AAGI colours to create colour scales suitable for use in
\CRANpkg{ggplot2} objects.
}
\seealso{
\link{aagi_palettes} \link{scale_fill_aagi}
}
