% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{scale_colour_aagi}
\alias{scale_colour_aagi}
\title{Construct AAGI Colour Scales for ggplot2}
\usage{
scale_colour_aagi(colourway = NULL, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{colourway}{\code{Character} name of a \acronym{AAGI} colourway in
\code{aagicolourways}.  Defaults to \code{NULL} and a random selection is made from
the three colourways, red, blue or green.}

\item{discrete}{\code{Boolean} argument indicating whether colour aesthetic is
discrete, \emph{e.g.}, a \code{factor} (\code{TRUE}) or continuous, *e.g., \code{numeric}
(\code{FALSE}).}

\item{reverse}{\code{Boolean} argument indicating whether the palette should be
reversed.  Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{discrete_scale()} or
\code{scale_colour_gradientn()}, used respectively when discrete is
\code{TRUE} or \code{FALSE}.}
}
\description{
Construct AAGI Colour Scales for {ggplot2}
}
