% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aagi_palettes.R
\name{scale_fill_aagi}
\alias{scale_fill_aagi}
\title{Construct AAGI Colour Fill Scales for ggplot2}
\usage{
scale_fill_aagi(palette = "colourful", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{\code{Character} name of a \acronym{AAGI} palette in
\code{\link[=aagi_palettes]{aagi_palettes()}}.  Defaults to \code{colourful}.}

\item{discrete}{\code{Boolean} argument indicating whether colour aesthetic is
discrete, \emph{e.g.}, a \code{factor} (\code{TRUE}) or continuous, *e.g., \code{numeric}
(\code{FALSE}).}

\item{reverse}{\code{Boolean} indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link[=discrete_scale]{discrete_scale()}} or
\code{\link[=scale_fill_gradientn]{scale_fill_gradientn()}}, used respectively when discrete is
\code{TRUE} or \code{FALSE}}
}
\description{
Construct AAGI Colour Fill Scales for ggplot2
}
\seealso{
\code{\link[=aagi_palettes]{aagi_palettes()}}
}
