% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aagi.R
\name{theme_aagi}
\alias{theme_aagi}
\title{A ggplot2 Theme Using a Unified AAGI Style and Typography for Plots}
\usage{
theme_aagi(
  base_size = 12,
  major_grid = FALSE,
  minor_grid = FALSE,
  border = FALSE,
  showtext_auto = TRUE,
  ...
)
}
\arguments{
\item{base_size}{\code{Numeric}. The figure's base size, set to 12 by default.
For presentations using a larger \code{base_size} will make the fonts larger and
more suitable for presentation slides, \emph{e.g.}, \code{base_size = 24}.}

\item{major_grid}{\code{Boolean}. Include major gridlines in the panel. Defaults
to \code{FALSE}.}

\item{minor_grid}{\code{Boolean}. Include minor gridlines in the panel. Defaults
to \code{FALSE}.}

\item{border}{\code{Boolean}. Include a border around the figure on all four
sides. Defaults to \code{FALSE}.}

\item{showtext_auto}{\code{Boolean}. If \code{TRUE}, enable \code{showtext::showtext_auto()}
that provides full \acronym{AAGI} font support. Defaults to \code{TRUE}.

\strong{Important:} \code{showtext::showtext_auto()} changes global rendering
behaviour for the entire R session/device. This can affect subsequent plots
(including non-\pkg{AAGIThemes} plots) until it is disabled. However, since
it is anticipated that this theme will only ever be used for official
\acronym{AAGI} work, this seems unlikely to be an issue.

To undo, call \code{\link[=showtext_aagi_off]{showtext_aagi_off()}}.}

\item{...}{Other arguments as passed along to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\description{
This theme follows a standard \acronym{AAGI} style including typography
guidelines that uses (hopefully) sensible defaults for \CRANpkg{ggplot2}
graphics.
}
\examples{
library("ggplot2")

p <- ggplot(mtcars) +
  geom_point(aes(wt, mpg)) +
  theme_aagi(showtext_auto = TRUE)

p

# Disable global showtext auto afterwards (recommended)
showtext_aagi_off()

}
\author{
Adam Sparks, \email{adam.sparks@curtin.edu.au}
}
