% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gt_aagi.R
\name{theme_gt_aagi}
\alias{theme_gt_aagi}
\title{Apply the AAGI Theme to a gt Object}
\usage{
theme_gt_aagi(x)
}
\arguments{
\item{x}{a \CRANpkg{gt} object}
}
\value{
a formatted \CRANpkg{gt} object
}
\description{
Apply the \acronym{AAGI} theme to a \CRANpkg{gt}.  An \acronym{AAGI}
formatted table body is grey with a teal header and white header font.
Header text is bold, text columns are left aligned, other columns are right
aligned.
}
\examples{
library(gt)
library(dplyr)
gt <- head(airquality) |>
  mutate(`Month Name` = "May") |>
  gt()
gt <- theme_gt_aagi(gt)
gt
}
\seealso{
Other tables: 
\code{\link{theme_ft_aagi}()}
}
\concept{tables}
